+++ src/features/data-upload/Uploader.module.css
@@ -2,9 +2,75 @@
   display: flex;
   flex-direction: column;
 }
-
 .upload{
   display:block;
   align-self: center;
-  margin-top: 50px;
+  margin-top: 30px;
+}
+
+.zone{
+    align-items: 'center';
+    border: 2px dashed #CCC;
+    border-radius: 20;
+    display: 'flex';
+    flex-direction: 'column';
+    height: '100%';
+    justify-content: 'center';
+    padding: 20;
+}
+
+.file {
+  background: linear-gradient(to bottom, #EEE, #DDD);
+  border-radius: 20px;
+  display: flex;
+  height: 120px;
+  width: 120px;
+  position: relative;
+  z-index: 10;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.info {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  padding-left: 10px;
+  padding-right: 10px;
+}
+
+.size {
+  background-color: var(--grey-light);
+  border-radius: 3px;
+  margin-bottom: 0.5em;
+  justify-content: center;
+  display: flex;
+}
+
+.name {
+  background-color: var(--grey-light);
+  border-radius: 3px;
+  font-size: 12px;
+  margin-bottom: 0.5em;
+}
+
+.progressBar {
+  bottom: 14px;
+  position: absolute;
+  width: 100%;
+  padding-left: 10px;
+  padding-right: 10px;
+}
+
+.zoneHover {
+  border-color: var(--grey-dim);
+}
+
+
+.remove {
+  height: 23px;
+  position: absolute;
+  right: 6px;
+  top: 6px;
+  width: 23px;
 }
